+++ modelicalang/__init__.py
@@ -1,8 +1,11 @@
+from __future__ import annotations
+
 __all__ = (
     "ModelicaVersion",
     "get_syntax_type",
     "ParsingExpressionLike",
     "returns_parsing_expression",
+    "ModelicaLangInternalWarning",
     "v3_4",
     "v3_5",
     "latest",
@@ -12,7 +15,10 @@ import enum
 from functools import lru_cache
 from typing import TYPE_CHECKING, Dict, Optional, Type, Union
 
-from typing_extensions import TypeAlias
+if TYPE_CHECKING:
+    from typing_extensions import TypeAlias
+
+from .exceptions import ModelicaLangInternalWarning
 
 if TYPE_CHECKING:
     from arpeggio import _ParsingExpressionLike  # noqa: F401