+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0a1] - 2023-07-12
+
+### Added
+
+- `modelicalang.ModelicaLangInternalWarning` Warning class
+    - It will be raised if arpeggio extension not enabled
+
+### Removed
+
+- Package dependency to `typing-extensions`
+
 ## [0.1.0a0] - 2023-02-01
 
 ### Added
@@ -19,5 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `modelicalang.v3_5.Syntax` [Modelica Concrete Syntax v3.5](https://specification.modelica.org/maint/3.5/modelica-concrete-syntax.html)
 
 
-[Unreleased]: https://github.com/ijknabla/ModelicaLanguageForPython/compare/v0.1.0a0...HEAD
+[Unreleased]: https://github.com/ijknabla/ModelicaLanguageForPython/compare/v0.1.0a1...HEAD
+[0.1.0a1]: https://github.com/ijknabla/ModelicaLanguageForPython/tree/v0.1.0a1
 [0.1.0a0]: https://github.com/ijknabla/ModelicaLanguageForPython/tree/v0.1.0a0